requirements.txt
Django==2.2.1
pytz==2019.1
scrypt==0.8.13
sqlparse==0.3.0
base58==2.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.1
django-ipfs-storage==0.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
ipfs-api==0.2.3
ipfsapi==0.4.4
ipfshttpclient==0.4.12
multiaddr==0.0.9
netaddr==0.7.19
pytz==2019.1
requests==2.23.0
scrypt==0.8.13
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
varint==1.0.2
base58==2.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.1
django-ipfs-storage==0.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
ipfs-api==0.2.3
ipfsapi==0.4.4
ipfshttpclient==0.4.12
multiaddr==0.0.9
netaddr==0.7.19
psycopg2==2.8.4
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.23.0
scrypt==0.8.13
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
varint==1.0.2
